refactor: 4회차 미션

+++ spring-hello/src/main/java/com/springhello/domain/product/entity/Product.java
@@ -8,16 +8,16 @@ public class Product {
 
     private final Long id;
     private final String name;
-    private final Long price;
+    private final Integer price;
 
     @Builder
-    private Product(Long id, String name, Long price) {
+    private Product(Long id, String name, Integer price) {
         this.id = id;
         this.name = name;
         this.price = price;
     }
 
-    public static Product createProduct(Long id, String name, Long price) {
+    public static Product createProduct(Long id, String name, Integer price) {
         return Product.builder()
                 .id(id)
                 .name(name)
